+++ src/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from "./artifact";
 export * from "./misc";
 export * from "./ethers";
 export * from "./task_helpers";
